Thread owner check already done on DBI level.
You're right. I didn't check this before and I didn't check whether this
feature is in DBI::PurePerl (grepping for 'thread' in lib/DBI/PurePerl.pm
didn't find anything, but I didn't review the code). I also miss a test
which proves that it works.

For this release, the patch which takes the p_thread stuff out of DBD::File
is attached. After it, some action seems required ...

Jens

git-svn-id: https://svn.perl.org/modules/dbi/trunk@13933 50811bd7-b8ce-0310-adc1-d9db26280581