Fixed parameter-count check for complicated statements using DBI::DBD::SqlEngine
When the params are in SQL::Statement::Param form, you cannot just count
the entries in the list,  but need to access the num method on the first
(and only) entry in the list.

Now this fix only covered a test case where the number of params was NIL
and I don't know if it is possible that the count is higher,  and if so,
if the test is still correct in those cases.

This fix needs tests!

See http://www.perlmonks.org/?node_id=910793 for an in-detail discussion

Signed-off-by: H.Merijn Brand (Tux) <hmbrand@cpan.org>

git-svn-id: https://svn.perl.org/modules/dbi/trunk@14891 50811bd7-b8ce-0310-adc1-d9db26280581