[PATCH 1/8] move method cache into dbi_ima_t struct (David Mitchell)
Date: Mon, 20 Feb 2012 18:38:13 +0000
Subject: [PATCH 1/8] move method cache into dbi_ima_t struct

The recently-added method cache stored the cache in magic attached to the
outer CV, rather than in the dbi_ima_t structure also attached to the CV,
because the dbi_ima_t structure couldn't be duped or freed, and so
couldn't have reference-counted SVs stored in it.

Change it so that the cache is now stored in the dbi_ima_t struct; but
still attach magic, who's sole function is to provide dup() and free()
methods which administer the dbi_ima_t struct.

This makes the code simpler, and also means that at method lookup time,
we don't have to retrieve any structures from magic.

Note that with this change, every DBI_dispatch method has a dbi_ima_t
struct, not just those with attributes. This will allow us to make further
use of it in the future.

There's a wart however. In thread cloning in perls before 5.8.9,
the any_ptr is copied from the old to the new CV *after* our magic dup
function is called, thus stopping us from being able to duplicate the
dbi_ima_t structure. We work round this (in code clearly delineated with
the BROKEN_DUP_ANY_PTR macro) by instead storing the current perl
interpreter address in in the dbi_ima_t; then in XS_DBI_dispatch,
we duplcate the dbi_ima_t at that point if the my_perl's don't match.
Clumsy, but should still be more efficient than the previous method of
storing the cache.

Once support for 5.8.x is dropped, it will be easy to remove this extra
code.


git-svn-id: https://svn.perl.org/modules/dbi/trunk@15194 50811bd7-b8ce-0310-adc1-d9db26280581