Subject: [PATCH] under ithreads, make DBIS efficient for DBD::* From: David Mitchell <davem@iabyn.com> Date: Fri, 10 Feb 2012 13:11:13 +0000
rather than the slow looking up of $DBI::_dbistate on every use of DBIS,
convert DBIS into a call to a C-level function that returns the address of
the dbi_state struct.

Since the C-level function is only directly callable from DBI, store its
address in an XSUB, so that the DBD:* modules can retrieve the function's
address and cache it in a static var.

We continue to store the dbi_state struct address within $DBI::_dbistate
too, so that DBD modules compiled against an older DBI will continue to
work if the DBI is upgraded but the DBD not recompiled.



git-svn-id: https://svn.perl.org/modules/dbi/trunk@15268 50811bd7-b8ce-0310-adc1-d9db26280581