Subject: [PATCH] cache imp_xxh in mg_ptr From: David Mitchell <davem@iabyn.com> Date: Wed, 18 Apr 2012 10:25:12 +0100
Currently, the imp_xxh_t structure is embedded in the PVX of an SV.
This SV is pointed to by mg->obj.

Make it so that a direct pointer to the imp_xxh is also cached in the
mg_ptr field (with mg_len set to 0 so that it won't be freed), thus
skipping a level or two of indirection when retrieving it.

At the same time, make dbih_getcom() short-circuit the common case, only
declaring dTHX and calling out to dbih_getcom2() in "unusual"
circumstances.

Together, this makes things infinitesimally quicker, especially on
threaded builds.


git-svn-id: https://svn.perl.org/modules/dbi/trunk@15269 50811bd7-b8ce-0310-adc1-d9db26280581