Ignore space issues of JSON formatting
The tests are about quotation of the value. Strip all leading/trailing
whitespace that is not part of the test,  treating '[99]' identical to
' [ 99 ] ' or '[99   ]'

reason is that the layout is not part of the test,  and formatters are
free to add/remove leading/trailing whitespace

These tests are only run if JSON::XS is installed.

git-svn-id: https://svn.perl.org/modules/dbi/trunk@15588 50811bd7-b8ce-0310-adc1-d9db26280581