Use a new attribute: f_dir_ext
Extended dir search now more explicit. One of the reasons not to use
a ref in f_dir is that the current definition is (very) clear and we
should not break that. What e.g. would we use to write new files?

If a table is opened from folder 2, then deleted and recreated, does
that mean we should re-use folder 2? If f_dir working as before this
problem disappears