Allow objects to be used as passwords without throwing an error.
I want to secure my passwords from being dumped in things like
stack traces by putting them in blessed objects with string overloading
so that only modules of my choosing, like DBI can actually read the
actual contents.

Unfortunately this is not possible because the DBI perl code
dies if the password argument is a ref. This patch changes this logic
to not die if the ref is blessed.

Apparently this restriction on password type is untested, so this patch
adds tests for both blessed and unblessed refs.