Replace "use vars" with "our"
Use of this pragma is discouraged in favour of "our" under 5.06+.

Also move strict/warnings higher in DBI.pm to cover the two big BEGIN
blocks, this caused a few other tweaks to code to compile under strict.

In a couple of tests vars was changed to my as package scope was
obviously not needed.