Remove "perl_" prefix from "perl_require_pv"
As per the docs, the perl_require_pv form is deprecated.

Note I think the second of these calls is actually a bug since
profile_class is a package name with colons, not a filename with
slashes. We're not currently checking ERRSV after this call so it's
probably silently failing. But such a fix is out of scope of this
commit.

Also it's probably worth moving to load_module instead as per the docs:

  It is analogous to the Perl code eval "require '$file'". It's even
  implemented that way; consider using load_module instead.